refactor embeds out of media-or-link
